feat: 포인트 조회 API 구현 및 단위 테스트 완료

+++ src/main/java/com/commercepaymentsystem/domain/point/dto/PointResponse.java
@@ -1,6 +1,6 @@
 package com.commercepaymentsystem.domain.point.dto;
 
 public record PointResponse(
-	Integer pointBalance
+	Long pointBalance
 ) {
 }
